Create validator for Dates: year, month, day
During the journey when a date is requested for the user, we always ask for the
year, month and day. These three attributes are mandatory in order to proceed to
the next question.

In the current implementation we are using a date helper that performs the
validation; I have decided to extract this out into a validator as seems a bit
more aligned with the ActiveModel lifecycle.

This is also aligned with how the MVP was implemented for the email validator.