ADR to introduce PageObject via Siteprism for Testing
This application consists on a high number of sequential questions and answers.
The current implementation is making the tests a bit difficult to read and
maintain.

We would like to clearly express the steps that the user is taking like:

- The user landed on the gender page
- The user transitioned to the contract detail page
- The user filled in his email address

While we can do that with the current implementation, we would need to rely
on modules, shared steps, and many other things that are not really necessary.