Add controller test for StartDateController
Controller tests are of use on this application because they include business
logic that is important for everything to work as expected.

We can see how on the `create` method there are validations, filters, policy
objects being exercised, etc, and we need to make sure that the controller is
performing as expected.

Testing all the scenarios at the system/feature test level is too slow for the
purpose of TDD.