User ActiveRecord persistence for Applicant Attributes
This is an intermediary commit where I am replacing all the attributes stored in
the session with a Database save!.

For now, I am storing the `applicant_id` on the session as I am not sure if this
can be exposed as part of the URL for privacy reasons.

In any case, this seems like a good improvement in the right direction.