Remove serialization from URN
I used serialization to promote encapsulation and improve the domain when
handling URN but since we now need to add a value that depends on the applicant
the code started to look a bit more complicated.

I have decided to simplify it by replacing it with a factory method that will be
used from the application in a `before_create`.