Update association from Application to Progress
According to #73, the `ApplicationProgress` domain model is linked to the
`Application` and not to the `Applicant`. This makes sense, as there is only
progress once the `Application` has been created.

On this commit, I am just changing the association (FK) from applicant to
application.

I am destroying a column because we don't have any production data.