Use Application as the root entity for Application factories
I am trying to make the factories more expressive, so when we create a factory
that is in an initial state, we created precisely that:

```ruby
create :application, :with_initial_checks_completed
```

This reads a bit better and enables a single point for application creation. It
is also removing the `traits` at the `application_progress` level for creating
teachers and trainees.