Add pending test for restarting the Teacher route
If an application (teacher route) goes back to the first question, and then
restarts the journey, the first questions are not remembered, because they are
not persisted anywhere.

To make this work we would need to persist all the applicant answers. This
approach will also simplify the codebase from different angles.

With the current implementation the user would need to click on the `yes` button
on one of the questions in order to complete the journey again.